101/requirements.txt
beautifulsoup4==4.10.0
lxml==4.6.3
pandas==1.3.3
requests2==2.16.0
selenium==3.141.0
